@@ -1,6 +1,6 @@
 module github.com/buzzxu/boys
 
-go 1.21
+go 1.22
 
 require (
 	github.com/bwmarrin/snowflake v0.3.0
